✅[CHORE] : BattleHistoryResultConstants 파일 이동 및 적용
