+++ SparkleAppTests/BattleHistoryResultTests/Mock/MockURLProtocol.swift
@@ -11,6 +11,7 @@ final class MockURLProtocol: URLProtocol {
     }
     override func startLoading() {
         if let error = MockURLProtocol.error {
+            self.client?.urlProtocol(self, didFailWithError: error)
         } else {
             self.client?.urlProtocol(self, didLoad: MockURLProtocol.stubResponseData ?? Data())
         }
