--- Uni/Network/ServiceProtocol/GetServiceProtocol.swift
@@ -1,5 +1,0 @@
-import Combine
-
-protocol GetServiceProtocol {
-    func getService<T: Decodable>(from url: String, isUseHeader: Bool) -> AnyPublisher<T, ErrorType>
-}
